+++ backend/src/main/java/ru/taskurotta/backend/statistics/datalisteners/LoggerDataListener.java
@@ -15,7 +15,7 @@ public class LoggerDataListener implements DataListener {
     private final static Logger logger = LoggerFactory.getLogger(LoggerDataListener.class);
 
     @Override
-    public void handle(String name, String actorId, int count, double value, long time) {
+    public void handle(String name, String actorId, long count, double value, long time) {
         if (logger.isInfoEnabled()) {
             logger.info("METRICS: [{}]#[{}]: value = [{}], count = [{}] collected at [{}]", name, actorId, value, count, new Date(time));
         }
